+++ ts/libs/polaris-nx/src/generators/apply-slo-mapping/schema.d.ts
@@ -1,0 +1,9 @@
+export interface ApplySloMappingGeneratorSchema {
+    sloMappingTsPath: string;
+}
+
+export interface ApplySloMappingGeneratorNormalizedSchema extends ApplySloMappingGeneratorSchema {
+    fullSloMappingTsPath: string;
+    sloMappingPathWithoutExtension: string;
+    fullSloMappingPathWithoutExtension: string;
+}